Close #583: The bin. format of the saved state is unstable
With some proper run-time checks, more robust auto-dep. GCC build, and
a fresh reg.-test baseline, I can't reproduce the anomaly any more.

+ Missing error check in the post-commit Git hook

+++ src/Model/cfg.h
@@ -3,7 +3,7 @@
 
 namespace Model {
 
-	using BasicNumberType = double;
+	using BasicNumberType = float;
 
 }
